[FT-41] - Созданы отдельные экраны:   - `TasksScreen` (задачи)   - `FinanceScreen` (финансы)   - `SleepScreen` (сон)   - `WaterScreen` (трекер воды)   - `SportScreen` (спорт)   - `DietScreen` (диета)   - `ProfileScreen` (профиль) - Перемещены экраны в папку `presentation/screens/trackers/` - Добавлены импорты экранов в `HomeScreen` - В `HomeScreen` реализован `GridView` с карточками для навигации по экранам - Добавлена проверка `context.mounted` перед вызовом `Navigator.pushReplacementNamed` в `_logout`, чтобы избежать ошибки `Don't use 'BuildContext's across async gaps` - Добавлен `ProfileScreen` и переход на него по кнопке в `AppBar`
