+++ mobile/lib/presentation/screens/auth/logout_screen.dart
