+++ mobile/lib/presentation/screens/auth/welcome_screen.dart
