+++ mobile/lib/presentation/screens/trackers/diet_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class DietScreen extends StatelessWidget {
+  const DietScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Диаета')),
+      body: const Center(child: Text('Экран диаеты')),
+    );
+  }
+}
