+++ mobile/lib/presentation/screens/trackers/finance_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class FinanceScreen extends StatelessWidget {
+  const FinanceScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Финансы')),
+      body: const Center(child: Text('Экран финансов')),
+    );
+  }
+}
