+++ mobile/lib/presentation/screens/trackers/tasks_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class TasksScreen extends StatelessWidget {
+  const TasksScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Задачи')),
+      body: const Center(child: Text('Экран задач')),
+    );
+  }
+}
